requirements.txt
docker
dxf
gitpython